responding to second round of feedback
