+++ RUFAS/data_structures/events.py
@@ -10,7 +10,7 @@ crop that is presently growing in a field will be harvested.
 
 from RUFAS.routines.EEE.enums import TillageImplement
 from RUFAS.routines.field.crop.harvest_operations import HarvestOperation
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_types import ManureType
 from RUFAS.time import Time
