+++ RUFAS/data_structures/manure_types.py
