+++ RUFAS/routines/field/field/field.py
@@ -3,7 +3,7 @@ from math import exp
 from typing import Any, Dict, List, Optional, Sequence
 
 from RUFAS.current_day_conditions import CurrentDayConditions
-from RUFAS.data_structures.crop_soil_to_manure_connection import (
+from RUFAS.data_structures.manure_to_crop_soil_connection import (
     ManureEventNutrientRequest,
     ManureEventNutrientRequestResults,
 )
@@ -25,8 +25,8 @@ from RUFAS.data_structures.events import (
     TillageEvent,
 )
 from RUFAS.routines.field.soil.soil import Soil
-from RUFAS.data_structures.nutrient_request import NutrientRequest, NutrientRequestResults
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequest, NutrientRequestResults
+from RUFAS.data_structures.manure_types import ManureType
 from RUFAS.time import Time
 from RUFAS.units import MeasurementUnits
