+++ RUFAS/routines/manure/field_manure_supplier.py
@@ -1,5 +1,5 @@
-from RUFAS.data_structures.nutrient_request import NutrientRequest, NutrientRequestResults
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequest, NutrientRequestResults
+from RUFAS.data_structures.manure_types import ManureType
 
 """
 The Field Manure Supplier uses ratios of dry mass to wet mass, Nitrogen to dry mass, and Phosphorus to dry mass to
