+++ RUFAS/routines/manure/manure_nutrients/manure_nutrients.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass, fields
 from typing import Dict
 
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_types import ManureType
 from RUFAS.units import MeasurementUnits
