+++ tests/test_manure_module/test_field_manure_supplier.py
@@ -1,8 +1,8 @@
 import pytest
 
 from RUFAS.routines.manure.field_manure_supplier import FieldManureSupplier
-from RUFAS.data_structures.nutrient_request import NutrientRequest, NutrientRequestResults
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequest, NutrientRequestResults
+from RUFAS.data_structures.manure_types import ManureType
 
 
 @pytest.mark.parametrize(
