+++ tests/test_manure_module/test_manure_nutrient_manager.py
@@ -7,8 +7,8 @@ from pytest_mock import MockerFixture
 
 from RUFAS.routines.manure.manure_nutrients.manure_nutrient_manager import ManureNutrientManager
 from RUFAS.routines.manure.manure_nutrients.manure_nutrients import ManureNutrients
-from RUFAS.data_structures.nutrient_request import NutrientRequest, NutrientRequestResults
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequest, NutrientRequestResults
+from RUFAS.data_structures.manure_types import ManureType
 
 
 @pytest.mark.parametrize("manure_type", [ManureType.LIQUID, ManureType.SOLID])
