+++ tests/test_manure_module/test_manure_nutrients.py
@@ -4,7 +4,7 @@ import pytest
 from pytest import approx, mark
 
 from RUFAS.routines.manure.manure_nutrients.manure_nutrients import ManureNutrients
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_types import ManureType
 
 
 @mark.parametrize(
