+++ tests/test_manure_module/test_manure_types.py
@@ -1,4 +1,4 @@
-from RUFAS.routines.manure.manure_treatments.manure_types import ManureType
+from RUFAS.data_structures.manure_types import ManureType
 
 
 def test_enum_ManureType() -> None:
