+++ tests/test_manure_module/test_nutrient_request_results.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pytest
 from pytest import approx
 
-from RUFAS.data_structures.nutrient_request import NutrientRequestResults
+from RUFAS.data_structures.manure_to_crop_soil_connection import NutrientRequestResults
 
 
 def test_nutrient_request_results_default_init() -> None:
